requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pydantic-settings
pydantic-core
python-dotenv
sqlalchemy
psycopg2-binary
alembic